NXBT-3162: Enable/Disable HAZELCAST with environment variable

+++ base/supervisord-nexus-config-ha.sh
@@ -1,11 +1,12 @@
 #!/bin/sh -ex
 
-# Delete hazelcast db 
-if test -d /nexus-data/db; then
+if [ "$HAZELCAST" == "true" ]; then
+  # Delete hazelcast db 
+  if test -d /nexus-data/db; then
     rm -rf /nexus-data/db/config
     rm -rf /nexus-data/db/component
     rm -rf /nexus-data/db/security
+  fi
+  grep -q 'nexus.clustered=true' /nexus-data/etc/nexus.properties && exit 0
+  echo 'nexus.clustered=true'  >> /nexus-data/etc/nexus.properties
 fi
-
-grep -q 'nexus.clustered=true' /nexus-data/etc/nexus.properties && exit 0
-echo 'nexus.clustered=true'  >> /nexus-data/etc/nexus.properties
